RFID: Fix integer overflow, fix printf usage (#742)
* Fix comparison of constant 300 with expression of type 'uint8_t' is always false

rfid-worker.cpp:111 if(validate_counts > (5 * 60))

* Fix format string is not a string literal (potentially insecure)

+++ applications/lfrfid/helpers/rfid-worker.h
@@ -43,6 +43,6 @@ private:
     void sq_write();
     void sq_write_start_validate();
     void sq_write_validate();
-    uint8_t validate_counts;
+    uint16_t validate_counts;
     void sq_write_stop_validate();
 };
